♻️ refactor: align code for pass the norminette
