+++ map/default.cub
@@ -21,13 +21,13 @@ C 203,194,176
         1111111111111111111111111
         1000000000110000000000001
         1011000001110000000O00001
-        1001000000000000O00000N01
-111111111011000001110000000000001
+        100101D100D00000O00000N01
+111111111011001001110000000000001
 10000000001100000111011111D111111
 11110111111111011100000010001    
 11110111111111011101010010001    
 11000000110101011100000010001    
-10000000000000001100000010001    
+10000000000000001100000010001   
 10000000000000001101010010001    
 1100001111010101111101111000111  
 11110111 1110101 101111010001
